Fix running Gazebo with NVidia GPU + PRIME
This makes Gazebo use the NVidia GPU (if present) when PRIME is
activated.

Setting XDG_RUNTIME_DIR and mounting /run/user are always required when
PRIME is installed (/etc/prime-discrete is `on` or `on-demand`).

In addition, when on-demand (offscreen) rendering is active
(/etc/prime-discrete is `on-demand`), the following environment
variables have to be set outside of Docker and exposed to Docker:

    export __NV_PRIME_RENDER_OFFLOAD=1
    export __GLX_VENDOR_LIBRARY_NAME=nvidia

Fixes #1.